+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/libs/ChatLib.kt
@@ -96,9 +96,9 @@ object ChatLib {
      * @param chatLineIDs the id(s) to be cleared
      */
     @JvmStatic
-    fun clearChat(vararg chatLineIDs: Int) {
+    fun clearChat(vararg chatLineIDs: Long) {
         for (chatLineID in chatLineIDs)
-            Client.getChatGUI()?.deleteChatLine(chatLineID)
+            Client.getChatGUI()?.deleteChatLine(chatLineID.toInt())
     }
 
     /**
@@ -250,7 +250,7 @@ object ChatLib {
      * @param replacements the new message(s) to be put in place of the old one
      */
     @JvmStatic
-    fun editChat(chatLineId: Int, vararg replacements: Message) {
+    fun editChat(chatLineId: Long, vararg replacements: Message) {
         editChat(
                 {
                     message -> message.getChatLineId() == chatLineId
@@ -286,7 +286,7 @@ object ChatLib {
             val chatLine = chatLineIterator.next()
 
             val result = toReplace(
-                    Message(chatLine.chatComponent).setChatLineId(chatLine.chatLineID)
+                    Message(chatLine.chatComponent).setChatLineId(chatLine.chatLineID.toLong())
             )
 
             if (!result) {
@@ -296,9 +296,9 @@ object ChatLib {
             chatLineIterator.remove()
 
             replacements.map {
-                val lineId = if (it.getChatLineId() == -1) 0 else it.getChatLineId()
+                val lineId = if (it.getChatLineId() == -1L) 0L else it.getChatLineId()
 
-                ChatLine(chatLine.updatedCounter, it.getChatMessage(), lineId)
+                ChatLine(chatLine.updatedCounter, it.getChatMessage(), lineId.toInt())
             }.forEach {
                 chatLineIterator.add(it)
             }
