+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/objects/Sound.kt
@@ -195,7 +195,7 @@ class Sound(private val config: ScriptObjectMirror) {
      *
      * @param model the model
      */
-    fun setAttenuation(model: Int) = apply { sndSystem!!.setAttenuation(this.source, model) }
+    fun setAttenuation(model: Long) = apply { sndSystem!!.setAttenuation(this.source, model.toInt()) }
 
     /**
      * Plays/resumes the sound
